centralize timedelta indexing tests in timedeltas.test_indexing
